requirements.txt
anyio==4.3.0
Authlib==1.3.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
Flask==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pycparser==2.22
python-dotenv==1.0.1
requests==2.31.0
sniffio==1.3.1
urllib3==2.2.1
Werkzeug==3.0.2
